requirements.txt
playwright==1.44.0
playwright-stealth==1.0.6
httpx==0.27.0
curl_cffi==0.7.2
duckdb==0.10.3
pandas==2.2.2
rich==13.7.1
python-dotenv==1.0.1
tenacity==8.3.0
marimo==0.6.19
plotly==5.22.0
kaleido==0.2.1
